requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
whisperx
torch==2.5.1+cu121
openai
yt_dlp
pydub
requests
azure-storage-blob
